+++ HISTORY.rst
@@ -8,6 +8,7 @@ v0.4.0 (NEXT)
 * Added Dichotomous-Hill model to list of dichotomous models
 * Allowed global overrides (such as BMR settings) as an optional input to ``Session.add_default_models``
 * Updated test-logic outputs for individual tests (and added tests)
+* For continuous summary datasets, rename ``responses`` array to ``means``
 
 
 v0.3.0 (2016-12-05)